Move BaseWriter to utils
